third_party/WebKit/LayoutTests/crypto/subtle/ec-export-private-key-expected.txt

Test exporting a private RSA key.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a PKCS8 key...

Exporting the key as JWK...
PASS exportedJWK.kty is "EC"
PASS exportedJWK.crv is "P-384"
PASS exportedJWK.d is "pJLOj6kAhMIn4aMveXTTnp_2en6HBew0GbNftgdYK-vUYeCxUgrHbsLdTptj665x"
PASS exportedJWK.x is "5V_ubEnY1SP1znv5wEJc5P9lBwi33lz7CVkBUjl5p_BCYC2zCFRzU2mBO1w_Xvho"
PASS exportedJWK.y is "KPWcxdxQmJKpiNOKjiUZ3j0MT9D72wmT448YUGwXYGxeJCSSRvHOlJg6U2HFvpg-"
PASS exportedJWK.alg is undefined
PASS exportedJWK.ext is true
PASS exportedJWK.key_ops is ['sign']
PASS exportedJWK.use is undefined

Exporting the key as PKCS8...
PASS: exportedPkcs8 should be [3081b6020100301006072a8648ce3d020106052b8104002204819e30819b0201010430a492ce8fa90084c227e1a32f7974d39e9ff67a7e8705ec3419b35fb607582bebd461e0b1520ac76ec2dd4e9b63ebae71a16403620004e55fee6c49d8d523f5ce7bf9c0425ce4ff650708b7de5cfb095901523979a7f042602db30854735369813b5c3f5ef86828f59cc5dc509892a988d38a8e2519de3d0c4fd0fbdb0993e38f18506c17606c5e24249246f1ce94983a5361c5be983e] and was
PASS successfullyParsed is true

TEST COMPLETE
